PUSH-1.1.2: Update ft_*.c files

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: serraoui <serraoui@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/05 11:43:44 by serraoui          #+#    #+#             */
-/*   Updated: 2023/11/05 11:45:00 by serraoui         ###   ########.fr       */
+/*   Updated: 2023/11/06 14:05:40 by serraoui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int				i;
-	unsigned char	*str1;
-	unsigned char	*str2;
+	size_t				i;
+	unsigned char		*str1;
+	unsigned char		*str2;
 
 	str1 = (unsigned char *) s1;
 	str2 = (unsigned char *) s2;
